log in and out

+++ db/schema.rb
@@ -51,7 +51,7 @@ ActiveRecord::Schema.define(version: 20170924164250) do
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.integer "price_cents", default: 0, null: false
-    t.float "minimum_size", default: 0.0, null: false
+    t.integer "minimum_size", default: 0, null: false
   end
 
   create_table "dough_fillings", force: :cascade do |t|
